feat : food 기본틀 생성

+++ src/main/java/com/heady/headyback/meal/dto/FoodDto.java
@@ -1,11 +1,13 @@
 package com.heady.headyback.meal.dto;
 
+import java.math.BigDecimal;
+
 import com.heady.headyback.meal.domain.Food;
 
 public record FoodDto(
 		Long id,
 		String name,
-		Integer calories
+		BigDecimal calories
 ) {
 	public static FoodDto of(Food food) {
 		return new FoodDto(food.getId(), food.getName(), food.getCalories());
